refactor(cli): extract subcommands into modules
Split monolithic main.rs (~950 lines) into focused modules:
- cli.rs: clap type definitions (Cli, Commands, AgentAction, TaskAction, PresenceAction)
- agent.rs: agent subcommand handlers (run, discover, bundle)
- task.rs: task subcommand handlers (send, status, stream)
- presence.rs: presence subcommand handlers (announce, discover, peers)
- common.rs: shared helpers (parse_capabilities)
- main.rs: slim entry point (~30 lines)

All 26 CLI tests preserved and passing.
