+++ crates/logos-messaging-a2a-cli/src/cli.rs
@@ -1,0 +1,572 @@
+use clap::{Parser, Subcommand};
+use std::path::PathBuf;
+
+#[derive(Debug, Parser)]
+#[command(
+    name = "logos-messaging-a2a",
+    about = "A2A protocol over Waku decentralized transport"
+)]
+pub struct Cli {
+    /// nwaku REST API URL
+    #[arg(long, default_value = "http://localhost:8645", global = true)]
+    pub waku: String,
+
+    #[command(subcommand)]
+    pub command: Commands,
+}
+
+#[derive(Debug, Subcommand)]
+pub enum Commands {
+    /// Agent management
+    Agent {
+        #[command(subcommand)]
+        action: AgentAction,
+    },
+    /// Task management
+    Task {
+        #[command(subcommand)]
+        action: TaskAction,
+    },
+    /// Presence management (Waku presence broadcasts)
+    Presence {
+        #[command(subcommand)]
+        action: PresenceAction,
+    },
+}
+
+#[derive(Debug, Subcommand)]
+pub enum AgentAction {
+    /// Run an agent that processes incoming tasks
+    Run {
+        /// Agent name
+        #[arg(long)]
+        name: String,
+        /// Comma-separated capabilities
+        #[arg(long, default_value = "text")]
+        capabilities: String,
+        /// Enable X25519+ChaCha20-Poly1305 encryption
+        #[arg(long)]
+        encrypt: bool,
+        /// Path to a persistent identity keyfile (hex-encoded 32-byte signing key).
+        /// If the file does not exist, a new key is generated and saved.
+        #[arg(long)]
+        keyfile: Option<PathBuf>,
+    },
+    /// Discover agents on the network
+    Discover,
+    /// Print this agent's IntroBundle (for sharing out-of-band)
+    Bundle,
+}
+
+#[derive(Debug, Subcommand)]
+pub enum TaskAction {
+    /// Send a task to an agent
+    Send {
+        /// Recipient agent public key (hex)
+        #[arg(long)]
+        to: String,
+        /// Text message to send
+        #[arg(long)]
+        text: String,
+    },
+    /// Check task status / poll for response
+    Status {
+        /// Task ID (UUID)
+        #[arg(long)]
+        id: String,
+    },
+    /// Follow a task's streaming output
+    Stream {
+        /// Task ID (UUID) to follow
+        #[arg(long)]
+        id: String,
+        /// How long to wait for the stream to complete (seconds)
+        #[arg(long, default_value = "30")]
+        timeout: u64,
+    },
+}
+
+#[derive(Debug, Subcommand)]
+pub enum PresenceAction {
+    /// Announce this agent on the presence topic
+    Announce {
+        /// Agent name
+        #[arg(long)]
+        name: String,
+        /// Comma-separated capabilities
+        #[arg(long, default_value = "text")]
+        capabilities: String,
+        /// TTL in seconds
+        #[arg(long, default_value = "300")]
+        ttl: u64,
+        /// Keep re-announcing every ttl/2 seconds
+        #[arg(long)]
+        repeat: bool,
+        /// Generate encrypted identity (X25519+ChaCha20-Poly1305)
+        #[arg(long)]
+        encrypt: bool,
+    },
+    /// Listen for presence announcements
+    Discover {
+        /// Filter by capability
+        #[arg(long)]
+        capability: Option<String>,
+        /// Keep listening instead of one-shot
+        #[arg(long)]
+        watch: bool,
+        /// How long to listen in one-shot mode (seconds)
+        #[arg(long, default_value = "10")]
+        timeout: u64,
+    },
+    /// Discover and list unique peers (deduplicated)
+    Peers {
+        /// Filter by capability
+        #[arg(long)]
+        capability: Option<String>,
+        /// Keep listening instead of one-shot
+        #[arg(long)]
+        watch: bool,
+        /// How long to listen in one-shot mode (seconds)
+        #[arg(long, default_value = "10")]
+        timeout: u64,
+    },
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use clap::error::ErrorKind;
+
+    fn try_parse(args: &[&str]) -> Result<Cli, clap::Error> {
+        Cli::try_parse_from(args)
+    }
+
+    // ── Presence Announce ──
+
+    #[test]
+    fn presence_announce_requires_name() {
+        let err = try_parse(&["cli", "presence", "announce"]).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::MissingRequiredArgument);
+    }
+
+    #[test]
+    fn presence_announce_with_name() {
+        let cli = try_parse(&["cli", "presence", "announce", "--name", "echo"]).unwrap();
+        match cli.command {
+            Commands::Presence {
+                action:
+                    PresenceAction::Announce {
+                        name,
+                        capabilities,
+                        ttl,
+                        repeat,
+                        encrypt,
+                    },
+            } => {
+                assert_eq!(name, "echo");
+                assert_eq!(capabilities, "text");
+                assert_eq!(ttl, 300);
+                assert!(!repeat);
+                assert!(!encrypt);
+            }
+            _ => panic!("expected Presence Announce"),
+        }
+    }
+
+    #[test]
+    fn presence_announce_all_flags() {
+        let cli = try_parse(&[
+            "cli",
+            "presence",
+            "announce",
+            "--name",
+            "bot",
+            "--capabilities",
+            "text,code",
+            "--ttl",
+            "600",
+            "--repeat",
+            "--encrypt",
+        ])
+        .unwrap();
+        match cli.command {
+            Commands::Presence {
+                action:
+                    PresenceAction::Announce {
+                        name,
+                        capabilities,
+                        ttl,
+                        repeat,
+                        encrypt,
+                    },
+            } => {
+                assert_eq!(name, "bot");
+                assert_eq!(capabilities, "text,code");
+                assert_eq!(ttl, 600);
+                assert!(repeat);
+                assert!(encrypt);
+            }
+            _ => panic!("expected Presence Announce"),
+        }
+    }
+
+    // ── Presence Discover ──
+
+    #[test]
+    fn presence_discover_defaults() {
+        let cli = try_parse(&["cli", "presence", "discover"]).unwrap();
+        match cli.command {
+            Commands::Presence {
+                action:
+                    PresenceAction::Discover {
+                        capability,
+                        watch,
+                        timeout,
+                    },
+            } => {
+                assert!(capability.is_none());
+                assert!(!watch);
+                assert_eq!(timeout, 10);
+            }
+            _ => panic!("expected Presence Discover"),
+        }
+    }
+
+    #[test]
+    fn presence_discover_with_filters() {
+        let cli = try_parse(&[
+            "cli",
+            "presence",
+            "discover",
+            "--capability",
+            "code",
+            "--watch",
+            "--timeout",
+            "30",
+        ])
+        .unwrap();
+        match cli.command {
+            Commands::Presence {
+                action:
+                    PresenceAction::Discover {
+                        capability,
+                        watch,
+                        timeout,
+                    },
+            } => {
+                assert_eq!(capability.as_deref(), Some("code"));
+                assert!(watch);
+                assert_eq!(timeout, 30);
+            }
+            _ => panic!("expected Presence Discover"),
+        }
+    }
+
+    // ── Presence Peers ──
+
+    #[test]
+    fn presence_peers_defaults() {
+        let cli = try_parse(&["cli", "presence", "peers"]).unwrap();
+        match cli.command {
+            Commands::Presence {
+                action:
+                    PresenceAction::Peers {
+                        capability,
+                        watch,
+                        timeout,
+                    },
+            } => {
+                assert!(capability.is_none());
+                assert!(!watch);
+                assert_eq!(timeout, 10);
+            }
+            _ => panic!("expected Presence Peers"),
+        }
+    }
+
+    #[test]
+    fn presence_peers_with_filters() {
+        let cli = try_parse(&[
+            "cli",
+            "presence",
+            "peers",
+            "--capability",
+            "summarize",
+            "--timeout",
+            "20",
+        ])
+        .unwrap();
+        match cli.command {
+            Commands::Presence {
+                action:
+                    PresenceAction::Peers {
+                        capability,
+                        watch,
+                        timeout,
+                    },
+            } => {
+                assert_eq!(capability.as_deref(), Some("summarize"));
+                assert!(!watch);
+                assert_eq!(timeout, 20);
+            }
+            _ => panic!("expected Presence Peers"),
+        }
+    }
+
+    // ── Global --waku flag ──
+
+    #[test]
+    fn global_waku_flag_with_presence() {
+        let cli = try_parse(&[
+            "cli",
+            "--waku",
+            "http://custom:9090",
+            "presence",
+            "discover",
+        ])
+        .unwrap();
+        assert_eq!(cli.waku, "http://custom:9090");
+    }
+
+    // ── Existing commands still parse ──
+
+    #[test]
+    fn agent_run_still_parses() {
+        let cli = try_parse(&["cli", "agent", "run", "--name", "test"]).unwrap();
+        matches!(cli.command, Commands::Agent { .. });
+    }
+
+    #[test]
+    fn task_send_still_parses() {
+        let cli = try_parse(&["cli", "task", "send", "--to", "abc", "--text", "hello"]).unwrap();
+        matches!(cli.command, Commands::Task { .. });
+    }
+
+    // ── Default waku URL ──
+
+    #[test]
+    fn default_waku_url() {
+        let cli = try_parse(&["cli", "agent", "discover"]).unwrap();
+        assert_eq!(cli.waku, "http://localhost:8645");
+    }
+
+    // ── Agent Discover ──
+
+    #[test]
+    fn agent_discover_parses() {
+        let cli = try_parse(&["cli", "agent", "discover"]).unwrap();
+        match cli.command {
+            Commands::Agent {
+                action: AgentAction::Discover,
+            } => {}
+            _ => panic!("expected Agent Discover"),
+        }
+    }
+
+    // ── Agent Bundle ──
+
+    #[test]
+    fn agent_bundle_parses() {
+        let cli = try_parse(&["cli", "agent", "bundle"]).unwrap();
+        match cli.command {
+            Commands::Agent {
+                action: AgentAction::Bundle,
+            } => {}
+            _ => panic!("expected Agent Bundle"),
+        }
+    }
+
+    // ── Agent Run details ──
+
+    #[test]
+    fn agent_run_defaults() {
+        let cli = try_parse(&["cli", "agent", "run", "--name", "echo"]).unwrap();
+        match cli.command {
+            Commands::Agent {
+                action:
+                    AgentAction::Run {
+                        name,
+                        capabilities,
+                        encrypt,
+                        keyfile,
+                    },
+            } => {
+                assert_eq!(name, "echo");
+                assert_eq!(capabilities, "text");
+                assert!(!encrypt);
+                assert!(keyfile.is_none());
+            }
+            _ => panic!("expected Agent Run"),
+        }
+    }
+
+    #[test]
+    fn agent_run_with_encrypt() {
+        let cli = try_parse(&["cli", "agent", "run", "--name", "secure", "--encrypt"]).unwrap();
+        match cli.command {
+            Commands::Agent {
+                action: AgentAction::Run { encrypt, .. },
+            } => {
+                assert!(encrypt);
+            }
+            _ => panic!("expected Agent Run"),
+        }
+    }
+
+    #[test]
+    fn agent_run_with_keyfile() {
+        let cli = try_parse(&[
+            "cli",
+            "agent",
+            "run",
+            "--name",
+            "persistent",
+            "--keyfile",
+            "/tmp/agent.key",
+        ])
+        .unwrap();
+        match cli.command {
+            Commands::Agent {
+                action:
+                    AgentAction::Run {
+                        name,
+                        keyfile,
+                        encrypt,
+                        ..
+                    },
+            } => {
+                assert_eq!(name, "persistent");
+                assert_eq!(keyfile, Some(PathBuf::from("/tmp/agent.key")));
+                assert!(!encrypt);
+            }
+            _ => panic!("expected Agent Run"),
+        }
+    }
+
+    // ── Task Send details ──
+
+    #[test]
+    fn task_send_fields() {
+        let cli = try_parse(&[
+            "cli",
+            "task",
+            "send",
+            "--to",
+            "abcdef",
+            "--text",
+            "hello world",
+        ])
+        .unwrap();
+        match cli.command {
+            Commands::Task {
+                action: TaskAction::Send { to, text },
+            } => {
+                assert_eq!(to, "abcdef");
+                assert_eq!(text, "hello world");
+            }
+            _ => panic!("expected Task Send"),
+        }
+    }
+
+    #[test]
+    fn task_send_missing_to() {
+        let err = try_parse(&["cli", "task", "send", "--text", "hi"]).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::MissingRequiredArgument);
+    }
+
+    #[test]
+    fn task_send_missing_text() {
+        let err = try_parse(&["cli", "task", "send", "--to", "abc"]).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::MissingRequiredArgument);
+    }
+
+    // ── Task Status ──
+
+    #[test]
+    fn task_status_parses() {
+        let cli = try_parse(&[
+            "cli",
+            "task",
+            "status",
+            "--id",
+            "550e8400-e29b-41d4-a716-446655440000",
+        ])
+        .unwrap();
+        match cli.command {
+            Commands::Task {
+                action: TaskAction::Status { id },
+            } => {
+                assert_eq!(id, "550e8400-e29b-41d4-a716-446655440000");
+            }
+            _ => panic!("expected Task Status"),
+        }
+    }
+
+    #[test]
+    fn task_status_missing_id() {
+        let err = try_parse(&["cli", "task", "status"]).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::MissingRequiredArgument);
+    }
+
+    // ── Missing subcommand ──
+
+    #[test]
+    fn missing_subcommand() {
+        let err = try_parse(&["cli"]).unwrap_err();
+        assert_eq!(
+            err.kind(),
+            ErrorKind::DisplayHelpOnMissingArgumentOrSubcommand
+        );
+    }
+
+    #[test]
+    fn unknown_flag_rejected() {
+        let err = try_parse(&["cli", "--bogus"]).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::UnknownArgument);
+    }
+
+    // ── Task Stream ──
+
+    #[test]
+    fn task_stream_with_id() {
+        let cli = try_parse(&["cli", "task", "stream", "--id", "task-42"]).unwrap();
+        match cli.command {
+            Commands::Task {
+                action: TaskAction::Stream { id, timeout },
+            } => {
+                assert_eq!(id, "task-42");
+                assert_eq!(timeout, 30); // default
+            }
+            _ => panic!("expected Task Stream"),
+        }
+    }
+
+    #[test]
+    fn task_stream_with_timeout() {
+        let cli = try_parse(&[
+            "cli",
+            "task",
+            "stream",
+            "--id",
+            "task-42",
+            "--timeout",
+            "60",
+        ])
+        .unwrap();
+        match cli.command {
+            Commands::Task {
+                action: TaskAction::Stream { id, timeout },
+            } => {
+                assert_eq!(id, "task-42");
+                assert_eq!(timeout, 60);
+            }
+            _ => panic!("expected Task Stream"),
+        }
+    }
+
+    #[test]
+    fn task_stream_missing_id() {
+        let err = try_parse(&["cli", "task", "stream"]).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::MissingRequiredArgument);
+    }
+}
