+++ crates/logos-messaging-a2a-cli/src/common.rs
@@ -1,0 +1,6 @@
+pub fn parse_capabilities(capabilities: &str) -> Vec<String> {
+    capabilities
+        .split(',')
+        .map(|s| s.trim().to_string())
+        .collect()
+}
